streamlit-faq-chatbot/requirements.txt
streamlit>=1.35
sentence-transformers>=2.7
faiss-cpu>=1.8
pandas>=2.2
nltk>=3.9
contractions>=0.1
rapidfuzz>=3.6
